refactor transfer ops builder
